requirements.txt
Flask
Werkzeug
PyJWT
Pillow
ImageHash
torch
transformers
google-cloud-vision
numpy
python-dotenv
flask-cors
